supported Unity 6000.0.27f1 and type inference.

+++ Assets/Editor/Build/Builder/FirebaseEditorBuilder.cs
@@ -1,32 +1,32 @@
-﻿using UnityEngine;
-using UnityPlugin.Core.Configure.Sns;
-using UnityEditor.iOS.Xcode;
-using System.IO;
-using System.Diagnostics;
-using System.Collections;
-namespace Editor.Build {
-public class FirebaseEditorBuilder : BaseEditorBuilder {
-    public const int BUILDER_ID = 6;
-    protected override void BuildiOS() {
-        if (false == this.pathDictionary.ContainsKey("fromFrameworkRoot") || false == this.pathDictionary.ContainsKey("destFrameworkRoot")) {
-            return;
-        }
-        string fromFrameworkRootPath = this.pathDictionary["fromFrameworkRoot"];
-        string destFrameworkRootPath = this.pathDictionary["destFrameworkRootAbsolute"];
-        string[] innerFilePathList = new string[] {
-            "Firebase/Firebase.h",
-            "Firebase/module.modulemap",
-        };
-        foreach (string innerFilePath in innerFilePathList) {
-            string fromPath = Path.Combine(fromFrameworkRootPath, innerFilePath);
-            string destPath = Path.Combine(destFrameworkRootPath, innerFilePath);
-            if (false != File.Exists(destPath)) {
-                File.Delete(destPath);
+﻿using System.IO;
+
+namespace Editor.Build
+{
+    public class FirebaseEditorBuilder : BaseEditorBuilder
+    {
+        public const int BUILDER_ID = 6;
+
+        protected override void BuildiOS()
+        {
+            if (false == pathDictionary.ContainsKey("fromFrameworkRoot") ||
+                false == pathDictionary.ContainsKey("destFrameworkRoot")) return;
+            var fromFrameworkRootPath = pathDictionary["fromFrameworkRoot"];
+            var destFrameworkRootPath = pathDictionary["destFrameworkRootAbsolute"];
+            var innerFilePathList = new[]
+            {
+                "Firebase/Firebase.h",
+                "Firebase/module.modulemap"
+            };
+            foreach (var innerFilePath in innerFilePathList)
+            {
+                var fromPath = Path.Combine(fromFrameworkRootPath, innerFilePath);
+                var destPath = Path.Combine(destFrameworkRootPath, innerFilePath);
+                if (File.Exists(destPath)) File.Delete(destPath);
+                File.Copy(fromPath, destPath);
             }
-            File.Copy(fromPath, destPath);
+
+            project.AddBuildProperty(targetGUID, "SWIFT_INCLUDE_PATHS",
+                "$(PROJECT_DIR)/Frameworks/Plugins/iOS/Frameworks/Firebase");
         }
-        project.AddBuildProperty(targetGUID, "SWIFT_INCLUDE_PATHS", "$(PROJECT_DIR)/Frameworks/Plugins/iOS/Frameworks/Firebase");
-        return;
     }
-}
 }
